requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
yfinance==1.2.0
pandas==2.2.2
google-genai==0.5.0
python-dotenv==1.0.1
